+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Fixed decryption with Active Storage 6 and `attachment.open`
+
 ## 0.3.5 (2020-04-13)
 
 - Added `array` type